requirements.txt
coverage==4.4.2
coveralls==1.2.0
Django==2.0.1
django-nose==1.4.5
djangorestframework==3.7.7
